file-storage-service/requirements.txt
Flask
minio
pyjwt==1.7.1